refactor(lotto): 도메인 역할별로 패키지 분리
- 도메인의 복잡성을 줄이고 역할을 명확히 표시하기 위해 패키지 분리
