+++ src/main/java/lotto/domain/core/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.core;
 
 import java.util.Arrays;
 import java.util.List;
