+++ src/main/java/lotto/domain/core/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.core;
 
 import java.util.Objects;
 import lotto.config.validation.FieldValidation;
