+++ src/test/java/lotto/application/in/LottoUseCaseTest.java
@@ -5,13 +5,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import lotto.Application;
 import lotto.config.context.ApplicationContext;
-import lotto.domain.BonusNumber;
-import lotto.domain.Lotto;
-import lotto.domain.LottoRank;
-import lotto.domain.LottoRound;
-import lotto.domain.LottoRoundResult;
-import lotto.domain.PurchaseAmount;
-import lotto.domain.WinningNumber;
+import lotto.domain.core.Lotto;
+import lotto.domain.core.LottoRank;
+import lotto.domain.input.BonusNumber;
+import lotto.domain.input.PurchaseAmount;
+import lotto.domain.input.WinningNumber;
+import lotto.domain.round.LottoRound;
+import lotto.domain.round.LottoRoundResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
