+++ src/test/java/lotto/domain/input/WinningNumberTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.input;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
