+++ src/test/java/lotto/domain/round/LottoRoundResultTest.java
@@ -1,6 +1,7 @@
-package lotto.domain;
+package lotto.domain.round;
 
 import java.util.List;
+import lotto.domain.core.LottoRank;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
